Udsp 2.3

+++ iq-provider-hbase/src/com/hex/bigdata/udsp/iq/provider/impl/model/HBasePage.java
@@ -9,10 +9,13 @@ import java.util.Map;
  */
 public class HBasePage {
     private List<Map<String, String>> records; // 每页记录集合
+
     private int pageSize = 10; // 每页显示条数
     private int pageIndex = 1; // 当前页
-    private int totalCount; // 总记录数
-    private int totalPage; // 总页数
+
+    private long totalCount; // 总记录数
+    private long totalPage; // 总页数
+
     private String startRow; // 开始rowkey
     private String stopRow; // 结束rowkey
 
@@ -20,18 +23,18 @@ public class HBasePage {
         super();
     }
 
-    public HBasePage(List<Map<String, String>> records, int pageSize, int pageIndex, int totalCount) {
+    public HBasePage(List<Map<String, String>> records, int pageSize, int pageIndex, long totalCount) {
         this.records = records;
         this.pageSize = pageSize;
         this.pageIndex = pageIndex;
         this.totalCount = totalCount;
     }
 
-    public int getTotalPage() {
+    public long getTotalPage() {
         return totalPage;
     }
 
-    public void setTotalPage(int totalPage) {
+    public void setTotalPage(long totalPage) {
         this.totalPage = totalPage;
     }
 
@@ -59,11 +62,11 @@ public class HBasePage {
         this.pageIndex = pageIndex;
     }
 
-    public int getTotalCount() {
+    public long getTotalCount() {
         return totalCount;
     }
 
-    public void setTotalCount(int totalCount) {
+    public void setTotalCount(long totalCount) {
         this.totalCount = totalCount;
         this.totalPage = totalCount % pageSize == 0 ? totalCount / pageSize
                 : (totalCount / pageSize + 1);
